pyproject.toml
[project]
name = "fastapiscrapeproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "celery>=5.5.2",
    "celery-redbeat>=2.3.2",
    "fastapi>=0.115.12",
    "greenlet>=3.2.2",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "redis>=6.1.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
